rework cli.py and headerdb complementer from the ground up
The cli rework is born from the need to to be able to process
the compdb general options, before the subcommands are processed.

For now the config kludge is ignored, it may be reincorporated later
or just rewritten in some other ways, time will tell.

The headerdb complementer was too complex and maybe not the right abstraction.
The new implementation just build a graph of the included-by
relationship between files.
The compile commands for headers are then generated dynamically
by traversing the graph in BFS order.